Added in javascript projects.
Had to clone the repository again because due to
some software I had downloaded there was an error
thrown back. Exact error is kex_exchange_identification
Software caused connection abort.

But it's likely because I had moved all of files
in the repo from one place to another without uh
cloning. Not sure if there's a simpler way to do
that, but I managed to finish it this time, so
hopefully this can push to origin.

As for the javascript projects, they are as per
the Odin description, so nothing to worry there.
DOM manip has not been done yet, through I intend
to do it in the coming weeks.