Added in calculator and editing css
There seems to be a way to add in the boxes with
flexbox with gap attr, and i want to experiment
with it, and for that I will be opening a new
branch to test this. just hope that this works.
because it doesnt seem to be working in git bash.